+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.4.0
+
+  * Removed first class Crosswalk (use it as a table instead)
+  * Added support for Monetize
+  * Renamed with-debug to debug
+
 ## 1.3.1
 
   * Moved github project from dirtyvagabond to Factual
@@ -9,10 +15,9 @@
   * Upgraded google-api-client to 1.8.0-beta
   * Upgraded clojure to 1.4.0
   * Upgraded slingshot to 0.10.2
-  * Upgraded data.json to 0.1.2 
+  * Upgraded data.json to 0.1.2
   * Added support for Facets
   * Added support for Geopulse
   * Added support for Reverse Geocoder
   * Added docs for using World Geographies table
   * Added debug mode
-